cleaner tests, restructuring
