+++ src/Fluent.Calculations.Primitives.Tests/ComplexValueType/Money/MoneyBuilder.cs
@@ -1,6 +1,5 @@
-﻿using Fluent.Calculations.Primitives.BaseTypes;
-
-namespace Fluent.Calculations.Primitives.Tests.Lab.Composition;
+﻿namespace Fluent.Calculations.Primitives.Tests.ComplexValueType;
+using Fluent.Calculations.Primitives.BaseTypes;
 
 public class MoneyBuilder
 {
