+++ src/Fluent.Calculations.Primitives.Tests/ComplexValueType/Money/ToMoney.cs
@@ -1,6 +1,6 @@
-﻿using Fluent.Calculations.Primitives.BaseTypes;
+﻿namespace Fluent.Calculations.Primitives.Tests.ComplexValueType;
+using Fluent.Calculations.Primitives.BaseTypes;
 using System.Runtime.CompilerServices;
-namespace Fluent.Calculations.Primitives.Tests.Lab.Composition;
 
 public static class ToMoney
 {
