+++ src/Fluent.Calculations.Primitives.Tests/DotNetGraph/CalculationGraphTests.cs
@@ -1,7 +1,7 @@
 using Fluent.Calculations.DotNetGraph;
 using Fluent.Calculations.Primitives.BaseTypes;
 
-namespace Fluent.Calculations.Primitives.Tests.Lab.DotNetGraph
+namespace Fluent.Calculations.Primitives.Tests.DotNetGraph
 {
     public class CalculationGraphTests
     {
